app/import/12_html_generator/cloud-run-files/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
anthropic>=0.25,<1
gunicorn